Refactor: Restructure Onboarding UI and Domain Layer
This commit refactors the onboarding feature's UI and domain layers for better organization and clarity.

Key changes:

- **Package Restructuring**:
    - `OnboardingViewModel` moved to `ui.onboarding.viewmodel`.
    - `OnboardingCardUiModel` moved to `ui.onboarding.model`.
    - `OnboardingUiState` and `Timing` moved to `ui.onboarding.state`.
    - `ColorUtils` and `DefaultPalette` (extracted from `OnboardingViewModel`) moved to `ui.onboarding.util`.
    - `OnboardingScreen` moved to `ui.onboarding.screen`.
- **Domain Layer Renaming**:
    - `EducationRepositoryInterface` renamed to `EducationRepository`.
    - `EducationRepository` (implementation) renamed to `EducationRepositoryImpl`.
- **Import Updates**:
    - All relevant files updated to reflect the new package structure and renamed interfaces/classes.
- **Minor Code Cleanup**:
