+++ app/src/main/java/dev/ansh/onboarding/onboarding/ui/onboarding/util/DefaultColorPalette.kt
@@ -1,0 +1,13 @@
+package dev.ansh.onboarding.onboarding.ui.onboarding.util
+
+import androidx.compose.ui.graphics.Color
+
+object DefaultPalette {
+    val ScreenGradientStart = Color(0xFF4A0F5C)
+    val ScreenGradientEnd = Color(0xFFAF1B87)
+    val CardBackground = Color(0xFF311B4B)
+    val StrokeStart = Color(0x66FFFFFF)
+    val StrokeEnd = Color(0x66FFFFFF)
+    val GradientStart = Color(0xFF6A1B9A)
+    val GradientEnd = Color(0xFFAD1457)
+}
